refactor(ui): extract scroll fade logic into reusable hook and update PrivacyClient and TermsClient to use it
